# docs/swagger.yaml
basePath: /swagger/visit-service
definitions:
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  model.Ward:
    properties:
      id:
        description: 病房ID
        type: string
      name:
        description: 病房名称
        type: string
      status:
        description: 病房状态(0:空闲,1:占用)
        type: integer
      type:
        description: 病房类型(0:普通,1:ICU)
        type: integer
    type: object
info:
  contact: {}
  description: visit-service API
  title: visit-service API
  version: "1.0"
paths:
...
  /ward:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Ward'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - 病房
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Ward'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Ward'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - 病房
  /ward/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Ward'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - 病房
  /ward/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - 病房
  /ward/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - 病房
  /ward/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Ward'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - 病房
...
swagger: "2.0"